requirements.txt
streamlit
pandas
joblib
xgboost
matplotlib
Pillow
